requirements.txt
asgiref==3.7.2
attrs==23.1.0
colorama==0.4.6
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-yasg==1.21.6
exceptiongroup==1.1.1
freezegun==1.2.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
